small changes to integrate category model
